Move HiDream O1 pipeline to modular diffusers
