--- docs/source/en/api/pipelines/hidream_o1.md
@@ -1,15 +1,0 @@
-# HiDream-O1
-
-HiDream-O1 is a Qwen3-VL based image generation model that predicts raw RGB image patches directly. Unlike HiDream-I1,
-it does not use a VAE component.
-
-The following model is available for the [`HiDreamO1ImagePipeline`] pipeline:
-
-| Model | Hugging Face Hub |
-|---|---|
-| HiDream-O1-Image | [`HiDream-ai/HiDream-O1-Image`](https://huggingface.co/HiDream-ai/HiDream-O1-Image) |
-| HiDream-O1-Image-Dev | [`HiDream-ai/HiDream-O1-Image-Dev`](https://huggingface.co/HiDream-ai/HiDream-O1-Image-Dev) |
-
-## HiDreamO1ImagePipeline
-
-[[autodoc]] HiDreamO1ImagePipeline
