+++ src/diffusers/__init__.py
@@ -464,6 +464,8 @@ else:
             "HeliosPyramidDistilledAutoBlocks",
             "HeliosPyramidDistilledModularPipeline",
             "HeliosPyramidModularPipeline",
+            "HiDreamO1AutoBlocks",
+            "HiDreamO1ModularPipeline",
             "HunyuanVideo15AutoBlocks",
             "HunyuanVideo15ModularPipeline",
             "LTXAutoBlocks",
@@ -566,7 +568,6 @@ else:
             "HeliosPipeline",
             "HeliosPyramidPipeline",
             "HiDreamImagePipeline",
-            "HiDreamO1ImagePipeline",
             "HunyuanDiTControlNetPipeline",
             "HunyuanDiTPAGPipeline",
             "HunyuanDiTPipeline",
@@ -1267,6 +1268,8 @@ if TYPE_CHECKING or DIFFUSERS_SLOW_IMPORT:
             HeliosPyramidDistilledAutoBlocks,
             HeliosPyramidDistilledModularPipeline,
             HeliosPyramidModularPipeline,
+            HiDreamO1AutoBlocks,
+            HiDreamO1ModularPipeline,
             HunyuanVideo15AutoBlocks,
             HunyuanVideo15ModularPipeline,
             LTXAutoBlocks,
@@ -1365,7 +1368,6 @@ if TYPE_CHECKING or DIFFUSERS_SLOW_IMPORT:
             HeliosPipeline,
             HeliosPyramidPipeline,
             HiDreamImagePipeline,
-            HiDreamO1ImagePipeline,
             HunyuanDiTControlNetPipeline,
             HunyuanDiTPAGPipeline,
             HunyuanDiTPipeline,
