+++ src/diffusers/modular_pipelines/modular_pipeline.py
@@ -130,6 +130,7 @@ MODULAR_PIPELINE_MAPPING = OrderedDict(
         ("qwenimage-edit", _create_default_map_fn("QwenImageEditModularPipeline")),
         ("qwenimage-edit-plus", _create_default_map_fn("QwenImageEditPlusModularPipeline")),
         ("qwenimage-layered", _create_default_map_fn("QwenImageLayeredModularPipeline")),
+        ("hidream-o1", _create_default_map_fn("HiDreamO1ModularPipeline")),
         ("z-image", _create_default_map_fn("ZImageModularPipeline")),
         ("helios", _create_default_map_fn("HeliosModularPipeline")),
         ("helios-pyramid", _helios_pyramid_map_fn),
